refactor(ui): simplify and optimize job list rendering - Remove unnecessary state management to improve performance - Consolidate job list components for reuse across views - Eliminate redundant NoHostView duplication across files
