+++ extensions/slurm/src/menu-bar.tsx
@@ -5,6 +5,7 @@ import { listJobsBrief, type Job } from "./lib/slurm";
 import { useActiveHosts, useSlurmUsers } from "./lib/session";
 import { fetchPerCluster } from "./lib/multi";
 import { openMasterInTerminal } from "./lib/ssh";
+import { countByState, formatMenuTitle } from "./lib/jobs";
 
 export default function MenuBar() {
   const { hosts, isLoading: hostsLoading } = useActiveHosts();
@@ -39,7 +40,7 @@ export default function MenuBar() {
   const clusters = results ?? [];
   const allJobs = clusters.flatMap((r) => (r.ok ? r.data : []));
   const counts = countByState(allJobs);
-  const title = formatTitle(counts);
+  const title = formatMenuTitle(counts);
   const anyError = clusters.some((r) => !r.ok);
   const tint = pickTint(counts, anyError);
   const hostsLabel = hosts.length > 0 ? hosts.join(", ") : "no cluster";
@@ -72,7 +73,7 @@ export default function MenuBar() {
         <MenuBarExtra.Section key={r.host} title={r.host}>
           {r.ok ? (
             <>
-              <MenuBarExtra.Item title={`Summary — ${formatTitle(countByState(r.data))}`} icon={Icon.BarChart} />
+              <MenuBarExtra.Item title={`Summary — ${formatMenuTitle(countByState(r.data))}`} icon={Icon.BarChart} />
               {renderJobsSubsection("Running", r.data, "RUNNING")}
               {renderJobsSubsection("Pending", r.data, "PENDING")}
               {renderJobsSubsection("Completing", r.data, "COMPLETING")}
@@ -140,24 +141,6 @@ function renderJobsSubsection(label: string, jobs: Job[], state: string) {
       ))}
     </>
   );
-}
-
-function countByState(jobs: Job[]): Record<string, number> {
-  const out: Record<string, number> = {};
-  for (const j of jobs) out[j.state] = (out[j.state] ?? 0) + 1;
-  return out;
-}
-
-function formatTitle(counts: Record<string, number>): string {
-  const r = counts.RUNNING ?? 0;
-  const p = counts.PENDING ?? 0;
-  const c = counts.COMPLETING ?? 0;
-  if (!r && !p && !c) return "idle";
-  const parts: string[] = [];
-  if (r) parts.push(`R${r}`);
-  if (p) parts.push(`P${p}`);
-  if (c) parts.push(`CG${c}`);
-  return parts.join("·");
 }
 
 function pickTint(counts: Record<string, number>, hasError: boolean): Color {
